+++ lib/src/utils/ansi_parser.dart
@@ -1,6 +1,4 @@
-import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 
 class AnsiParser {
   static const TEXT = 0, BRACKET = 1, CODE = 2;
@@ -123,11 +121,6 @@ class AnsiParser {
         style: TextStyle(
           color: foreground,
           backgroundColor: background,
-        ),
-        recognizer: LongPressGestureRecognizer()
-          ..onLongPress = () {
-            Clipboard.setData(ClipboardData(text: text));
-            // Toast.toast("Copy to paste board");
-          });
+        ));
   }
 }